# feat(bitnet): implement QKV projection for attention